todas as funcionalidade prontas com excessão da alteração de viagem

+++ app/src/main/java/com/example/gerenciaviagens/bean/Refeicoes.java
@@ -4,7 +4,7 @@ import android.os.Parcel;
 import android.os.Parcelable;
 
 public class Refeicoes implements Parcelable {
-    private int id;
+    private long id;
     private Viagem viagem;
     private float custo_refeicoes;
     private int refeicoes_dia;
@@ -33,11 +33,11 @@ public class Refeicoes implements Parcelable {
         }
     };
 
-    public int getId() {
+    public long getId() {
         return id;
     }
 
-    public void setId(int id) {
+    public void setId(long id) {
         this.id = id;
     }
 
@@ -80,7 +80,7 @@ public class Refeicoes implements Parcelable {
 
     @Override
     public void writeToParcel(Parcel dest, int flags) {
-        dest.writeInt(id);
+        dest.writeLong(id);
         dest.writeParcelable(viagem, flags);
         dest.writeFloat(custo_refeicoes);
         dest.writeInt(refeicoes_dia);
